COMPUTE-SLOT-POSITIONS is now defined on BINARY-LITTLE-ENDIAN-OBJECT
This and BINARY-BIG-ENDIAN-OBJECT are also created in this commit. These
classes make sure that the input to COMPUTE-SLOT-POSITIONS are objects
that make sense to this function. We must assume the object is a
BINARY-DATA-OBJECT as well as having something useful defined as to its
endian definition.

There is an alternate way to implement this... basically define an extra
generic function to handle the differing endian directions, but we are
passing on this option for the time being.